Move async classes to their own subfolder
