+++ mobx/lib/mobx.dart
@@ -42,9 +42,9 @@ export 'package:mobx/src/api/annotations.dart'
     show action, computed, observable, Store;
 export 'package:mobx/src/api/action.dart';
 export 'package:mobx/src/api/context.dart';
-export 'package:mobx/src/api/observable_future.dart'
+export 'package:mobx/src/api/async/observable_future.dart'
     show ObservableFuture, FutureStatus;
-export 'package:mobx/src/api/observable_stream.dart'
+export 'package:mobx/src/api/async/observable_stream.dart'
     show ObservableStream, StreamStatus;
 export 'package:mobx/src/api/observable_collections.dart'
     show
@@ -58,4 +58,4 @@ export 'package:mobx/src/api/observable_collections.dart'
         SetChange,
         SetChangeListener;
 export 'package:mobx/src/api/reaction.dart';
-export 'package:mobx/src/api/async_action.dart' show AsyncAction;
+export 'package:mobx/src/api/async/async_action.dart' show AsyncAction;
