+++ mobx/test/observable_stream_test.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import 'package:mobx/src/api/observable_stream.dart';
+import 'package:mobx/src/api/async/observable_stream.dart';
 import 'package:mobx/src/api/reaction.dart';
 import 'package:test/test.dart';
